Added support for STL containers in class parsing. The MapClass() routine of rb::data::Mapper can now handle the following STL containers: vector, set, list, deque, multiset (with any basic data for the template argument).
Support is also present for ReadData(), to parse the individual STL
elements's names into a vector of strings.

Note that STL support still is not present for FindBasicAddr, that is finding
the address of a member of an STL container for histogramming purposes. So for
now, histogramming an STL member would still require TTreeFormula.