Implementation of rb::ClassFormula. This class can evaluate expressions like TTreeFormula, but it uses the direct addresses of class members contained within the expression for formula evalution. This seems to give a pretty good performance advantage over TTreeFormula.
The class has been tested and seems to work okay. It is currently implemented as the "default"
formula type (in Event.cxx), but if it's parsing fails, the old methods are still used. Eventually,
all of the old stuff should be removed and all of the code to generate formulas should be cleaned up,
with only ClassFormula being used.